+++ src/Action/History/ActionHistory.php
@@ -6,13 +6,14 @@ use BristolSU\Support\Action\ActionInstance;
 use Database\Factories\ActionHistoryFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 /**
  * Represents the action history.
  */
 class ActionHistory extends Model
 {
-    use HasFactory;
+    use HasFactory, SoftDeletes;
 
     protected $table = 'action_histories';