+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Soft deletes on most models, and ability to restore the models and their children
+
 ## [5.0.2] - (23/10/2021)
 
 ### Changed